fix: use casts instead of .longValue for integer-based cPars
See under 5.3 in https://doc.omnetpp.org/omnetpp/api/APIChanges.html

This will also make RINASim compatible with OMNeT++ 6, where
.longValue() is replaced with .intValue().

+++ playground/DAF/AE/AESInt.cc
@@ -86,7 +86,7 @@ void AESInt::initialize()
 
 
     //MAX PDU Size
-    maxSize = par("maxPDUsize").longValue();
+    maxSize = par("maxPDUsize");
 
     send = 0;
     sendSize = 0;
